Refactoring of Avro-hive type conversion utils
