+++ gobblin-data-management/src/main/java/org/apache/gobblin/data/management/conversion/hive/query/HiveAvroORCQueryGenerator.java
@@ -23,9 +23,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 
-import org.apache.avro.AvroRuntimeException;
-import org.apache.avro.LogicalType;
-import org.apache.avro.LogicalTypes;
 import org.apache.avro.Schema;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
@@ -33,8 +30,6 @@ import org.apache.hadoop.hive.metastore.api.FieldSchema;
 import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.serde.serdeConstants;
 import org.apache.hadoop.hive.serde2.SerDeException;
-import org.apache.hadoop.hive.serde2.avro.AvroSerDe;
-import org.apache.hadoop.hive.serde2.avro.AvroSerdeException;
 import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspector;
 import org.apache.hadoop.hive.serde2.typeinfo.ListTypeInfo;
 import org.apache.hadoop.hive.serde2.typeinfo.MapTypeInfo;
@@ -62,6 +57,7 @@ import org.apache.gobblin.data.management.conversion.hive.task.HiveConverterUtil
 import org.apache.gobblin.util.HiveAvroTypeConstants;
 
 import static org.apache.gobblin.data.management.conversion.hive.entities.StageableTableMetadata.SCHEMA_SOURCE_OF_TRUTH;
+import static org.apache.gobblin.data.management.conversion.hive.utils.AvroHiveTypeUtils.generateAvroToHiveColumnMapping;
 import static org.apache.gobblin.util.AvroUtils.sanitizeSchemaString;
 
 
@@ -379,174 +375,7 @@ public class HiveAvroORCQueryGenerator {
    * @param topLevel If this is first level
    * @return Generate Hive columns with types for given Avro schema
    */
-  private static String generateAvroToHiveColumnMapping(Schema schema, Optional<Map<String, String>> hiveColumns,
-      boolean topLevel, String datasetName) {
-    if (topLevel && !schema.getType().equals(Schema.Type.RECORD)) {
-      throw new IllegalArgumentException(
-          String.format("Schema for table must be of type RECORD. Received type: %s for dataset %s", schema.getType(),
-              datasetName));
-    }
 
-    StringBuilder columns = new StringBuilder();
-    boolean isFirst;
-    switch (schema.getType()) {
-      case RECORD:
-        isFirst = true;
-        if (topLevel) {
-          for (Schema.Field field : schema.getFields()) {
-            if (isFirst) {
-              isFirst = false;
-            } else {
-              columns.append(", \n");
-            }
-            String type = generateAvroToHiveColumnMapping(field.schema(), hiveColumns, false, datasetName);
-            if (hiveColumns.isPresent()) {
-              hiveColumns.get().put(field.name(), type);
-            }
-            String flattenSource = field.getProp("flatten_source");
-            if (StringUtils.isBlank(flattenSource)) {
-              flattenSource = field.name();
-            }
-            columns.append(
-                String.format("  `%s` %s COMMENT 'from flatten_source %s'", field.name(), type, flattenSource));
-          }
-        } else {
-          columns.append(HiveAvroTypeConstants.AVRO_TO_HIVE_COLUMN_MAPPING_V_12.get(schema.getType())).append("<");
-          for (Schema.Field field : schema.getFields()) {
-            if (isFirst) {
-              isFirst = false;
-            } else {
-              columns.append(",");
-            }
-            String type = generateAvroToHiveColumnMapping(field.schema(), hiveColumns, false, datasetName);
-            columns.append("`").append(field.name()).append("`").append(":").append(type);
-          }
-          columns.append(">");
-        }
-        break;
-      case UNION:
-        Optional<Schema> optionalType = isOfOptionType(schema);
-        if (optionalType.isPresent()) {
-          Schema optionalTypeSchema = optionalType.get();
-          columns.append(generateAvroToHiveColumnMapping(optionalTypeSchema, hiveColumns, false, datasetName));
-        } else {
-          columns.append(HiveAvroTypeConstants.AVRO_TO_HIVE_COLUMN_MAPPING_V_12.get(schema.getType())).append("<");
-          isFirst = true;
-          for (Schema unionMember : schema.getTypes()) {
-            if (Schema.Type.NULL.equals(unionMember.getType())) {
-              continue;
-            }
-            if (isFirst) {
-              isFirst = false;
-            } else {
-              columns.append(",");
-            }
-            columns.append(generateAvroToHiveColumnMapping(unionMember, hiveColumns, false, datasetName));
-          }
-          columns.append(">");
-        }
-        break;
-      case MAP:
-        columns.append(HiveAvroTypeConstants.AVRO_TO_HIVE_COLUMN_MAPPING_V_12.get(schema.getType())).append("<");
-        columns.append("string,")
-            .append(generateAvroToHiveColumnMapping(schema.getValueType(), hiveColumns, false, datasetName));
-        columns.append(">");
-        break;
-      case ARRAY:
-        columns.append(HiveAvroTypeConstants.AVRO_TO_HIVE_COLUMN_MAPPING_V_12.get(schema.getType())).append("<");
-        columns.append(generateAvroToHiveColumnMapping(schema.getElementType(), hiveColumns, false, datasetName));
-        columns.append(">");
-        break;
-      case NULL:
-        break;
-      case BYTES:
-      case DOUBLE:
-      case ENUM:
-      case FIXED:
-      case FLOAT:
-      case INT:
-      case LONG:
-      case STRING:
-      case BOOLEAN:
-        // Handling Avro Logical Types which should always sit in leaf-level.
-        boolean isLogicalTypeSet = false;
-        try {
-          String hiveSpecificLogicalType = generateHiveSpecificLogicalType(schema);
-          if (StringUtils.isNoneEmpty(hiveSpecificLogicalType)) {
-            isLogicalTypeSet = true;
-            columns.append(hiveSpecificLogicalType);
-            break;
-          }
-        } catch (AvroSerdeException ae) {
-          log.error("Failed to generate logical type string for field" + schema.getName() + " due to:", ae);
-        }
-
-        LogicalType logicalType = LogicalTypes.fromSchemaIgnoreInvalid(schema);
-        if (logicalType != null) {
-          switch (logicalType.getName().toLowerCase()) {
-            case HiveAvroTypeConstants.DATE:
-              LogicalTypes.Date dateType = (LogicalTypes.Date) logicalType;
-              dateType.validate(schema);
-              columns.append("date");
-              isLogicalTypeSet = true;
-              break;
-            case HiveAvroTypeConstants.DECIMAL:
-              LogicalTypes.Decimal decimalType = (LogicalTypes.Decimal) logicalType;
-              decimalType.validate(schema);
-              columns.append(String.format("decimal(%s, %s)", decimalType.getPrecision(), decimalType.getScale()));
-              isLogicalTypeSet = true;
-              break;
-            case HiveAvroTypeConstants.TIME_MILLIS:
-              LogicalTypes.TimeMillis timeMillsType = (LogicalTypes.TimeMillis) logicalType;
-              timeMillsType.validate(schema);
-              columns.append("timestamp");
-              isLogicalTypeSet = true;
-              break;
-            default:
-              log.error("Unsupported logical type" + schema.getLogicalType().getName() +  ", fallback to physical type");
-          }
-        }
-
-        if (!isLogicalTypeSet) {
-          columns.append(HiveAvroTypeConstants.AVRO_TO_HIVE_COLUMN_MAPPING_V_12.get(schema.getType()));
-        }
-        break;
-      default:
-        String exceptionMessage =
-            String.format("DDL query generation failed for \"%s\" of dataset %s", schema, datasetName);
-        log.error(exceptionMessage);
-        throw new AvroRuntimeException(exceptionMessage);
-    }
-
-    return columns.toString();
-  }
-
-  /**
-   * Referencing org.apache.hadoop.hive.serde2.avro.SchemaToTypeInfo#generateTypeInfo(org.apache.avro.Schema) on
-   * how to deal with logical types that supported by Hive but not by Avro(e.g. VARCHAR).
-   *
-   * If unsupported logical types found, return empty string as a result.
-   * @param schema Avro schema
-   * @return
-   * @throws AvroSerdeException
-   */
-  public static String generateHiveSpecificLogicalType(Schema schema) throws AvroSerdeException {
-    // For bytes type, it can be mapped to decimal.
-    Schema.Type type = schema.getType();
-
-    if (type == Schema.Type.STRING && AvroSerDe.VARCHAR_TYPE_NAME
-        .equalsIgnoreCase(schema.getProp(AvroSerDe.AVRO_PROP_LOGICAL_TYPE))) {
-      int maxLength = 0;
-      try {
-        maxLength = schema.getJsonProp(AvroSerDe.AVRO_PROP_MAX_LENGTH).getValueAsInt();
-      } catch (Exception ex) {
-        throw new AvroSerdeException("Failed to obtain maxLength value from file schema: " + schema, ex);
-      }
-      return String.format("varchar(%s)", maxLength);
-    } else {
-      return StringUtils.EMPTY;
-    }
-  }
 
   /***
    * Use destination table schema to generate column mapping
@@ -641,37 +470,6 @@ public class HiveAvroORCQueryGenerator {
     } else {
       throw new RuntimeException("Unknown type encountered: " + type);
     }
-  }
-
-  /***
-   * Check if the Avro Schema is of type OPTION
-   * ie. [null, TYPE] or [TYPE, null]
-   * @param schema Avro Schema to check
-   * @return Optional Avro Typed data if schema is of type OPTION
-   */
-  private static Optional<Schema> isOfOptionType(Schema schema) {
-    Preconditions.checkNotNull(schema);
-
-    // If not of type UNION, cant be an OPTION
-    if (!Schema.Type.UNION.equals(schema.getType())) {
-      return Optional.<Schema>absent();
-    }
-
-    // If has more than two members, can't be an OPTION
-    List<Schema> types = schema.getTypes();
-    if (null != types && types.size() == 2) {
-      Schema first = types.get(0);
-      Schema second = types.get(1);
-
-      // One member should be of type NULL and other of non NULL type
-      if (Schema.Type.NULL.equals(first.getType()) && !Schema.Type.NULL.equals(second.getType())) {
-        return Optional.of(second);
-      } else if (!Schema.Type.NULL.equals(first.getType()) && Schema.Type.NULL.equals(second.getType())) {
-        return Optional.of(first);
-      }
-    }
-
-    return Optional.<Schema>absent();
   }
 
   /***
